requirements.txt
flask==3.0.0
pydantic==2.9.2
langchain==0.3.7
langchain-openai==0.2.9
langgraph==0.2.48
python-dotenv==1.0.0
requests==2.31.0